+++ src/server/app.js
@@ -4,11 +4,10 @@ import express from 'express';
 import spiderDetector from 'spider-detector';
 import tracer from 'tracer';
 import { renderToString, renderToStaticMarkup } from 'react-dom/server';
-import { Root } from '../app/components';
-import configureStore from '../app/configureStore';
+import { Root, Routes } from '../shared/components';
+import configureStore from '../shared/configureStore';
 import { createElement } from 'react';
 import { match, createMemoryHistory } from 'react-router';
-import Routes from '../app/components/Routes';
 import proxy from 'express-http-proxy';
 import { DB_NAME, DB_HOST } from './config';
 import { encode } from '../shared/encoding';
