+++ src/shared/components/Post.jsx
