+++ src/shared/components/index.js
@@ -3,3 +3,4 @@ export { default as App } from './App';
 export { default as Home } from './Home';
 export { default as Post } from './Post';
 export { default as Page } from './Page';
+export { default as Routes } from './Routes';
