+++ src/shared/reducers/index.js
